Moved IScriptManager out into new file.
