+++ src/core/Wyam.Configuration/Wyam.Configuration.csproj
@@ -233,6 +233,7 @@
     <Compile Include="Assemblies\AssemblyResolver.cs" />
     <Compile Include="Assemblies\AssemblyLoader.cs" />
     <Compile Include="ConfigScript\CacheManager.cs" />
+    <Compile Include="ConfigScript\IScriptManager.cs" />
     <Compile Include="ConfigScript\LiftingWalker.cs" />
     <Compile Include="ConfigScript\ScriptCompilationException.cs" />
     <Compile Include="Directives\AssemblyDirective.cs" />
